requirement.txt
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
boto3==1.23.8
botocore==1.26.8
brotlipy==0.7.0
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
click==8.0.4
cloudpickle==2.0.0
conda==4.9.2
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cos-python-sdk-v5==1.9.17
crcmod==1.7
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cycler==0.11.0
dask==2022.4.0
decorator==5.1.1
Deprecated==1.2.13
dicttoxml==1.7.4
einops==0.4.1
executing==0.8.3
filelock==3.6.0
fonttools==4.31.2
fsspec==2022.3.0
future==0.18.2
grpcio==1.43.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-resources==5.4.0
ipython==8.1.0
jedi==0.18.1
jmespath==1.0.0
jsonschema==4.4.0
kiwisolver==1.4.2
lmdb==1.3.0
locket==0.2.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
msgpack==1.0.3
networkx==2.7.1
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
parso==0.8.3
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome==3.14.1
Pygments==2.11.2
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
ray==1.11.0
redis==4.1.4
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
ruamel-yaml==0.15.87
s3transfer==0.5.2
scikit-image==0.14.2
scipy==1.8.0
Shapely==1.8.1.post1
six @ file:///tmp/build/80754af9/six_1605205327372/work
stack-data==0.2.0
tensorboardX==2.5
timm==0.5.4
toolz==0.11.2
torch==1.6.0+cu101
torchaudio==0.8.1
torchvision==0.7.0+cu101
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets==5.1.1
typing-extensions==4.1.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.7.0
